requirements.txt
# Core PyTorch ecosystem
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Machine learning
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and logging
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Progress bars and utilities
tqdm>=4.65.0
rich>=13.0.0

# Data handling
h5py>=3.8.0
zarr>=2.14.0

# Pre-trained models
transformers>=4.30.0
datasets>=2.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Development tools
jupyter>=1.0.0
ipython>=8.14.0

# Optional: GPU acceleration
# nvidia-ml-py>=11.0.0  # Uncomment for GPU monitoring

# Optional: Audio enhancement
# pesq>=0.0.3  # Uncomment for PESQ metric
# pystoi>=0.3.3  # Uncomment for STOI metric
